server/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydub==0.25.1
ffmpeg-python==0.2.0
openai-whisper==20231117
torch==2.4.1
six
unicodedata2
flask
flask-cors
requests
python-dotenv
